Found Max

+++ FindMaxUsingGenerics/Program.cs
@@ -31,5 +31,8 @@ Console.WriteLine(A);
 Console.WriteLine(B);
 Console.WriteLine(C);
 
-Find<int> maxFloat = new Find<int>();
+Find<double> maxFloat = new Find<double>();
 maxFloat.CompareFloat(A, B, C);
+
+
+
